requirements.txt
torch>=1.0.0
pandas
numpy
torchvision
scikit-learn
Pillow
